requirements.txt
Django==1.5.1
PyVCF==0.6.4
amqp==1.0.12
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.2.0
billiard==2.7.3.31
biopython==1.62b
celery==3.0.21
distribute==0.7.3
django-celery==3.0.17
django-nose==1.2
django-registration==1.0
ipython==0.13.2
jsonfield==0.9.17
kombu==2.5.12
nose==1.3.0
python-dateutil==2.1
pytz==2013b
six==1.3.0
sympy==0.7.3
wsgiref==0.1.2
boto==2.15.0
